requirements.txt
requests>=2.28.0
redis>=4.5.0
python-dotenv>=1.0.0
